Always recompile test contracts
