+++ crates/locutus-runtime/src/tests/mod.rs
@@ -10,6 +10,7 @@ use locutus_stdlib::prelude::{
 
 use crate::ContractStore;
 
+mod contract;
 mod time;
 
 static TEST_NO: AtomicUsize = AtomicUsize::new(0);
@@ -33,42 +34,35 @@ pub(crate) fn get_test_module(name: &str) -> Result<Vec<u8>, Box<dyn std::error:
         let path = dirs.nth(2).unwrap();
         path.join("tests").join(name.replace('_', "-"))
     };
-    let mut contract_build_path = module_path
-        .join("build/locutus")
+    const TARGET_DIR_VAR: &str = "CARGO_TARGET_DIR";
+    let target = std::env::var(TARGET_DIR_VAR).map_err(|_| "CARGO_TARGET_DIR should be set")?;
+    println!("trying to compile the test contract, target: {target}");
+    // attempt to compile it
+    const RUST_TARGET_ARGS: &[&str] = &["build", "--target"];
+    const WASI_TARGET: &str = "wasm32-wasi";
+    let cmd_args = RUST_TARGET_ARGS
+        .iter()
+        .copied()
+        .chain([WASI_TARGET])
+        .collect::<Vec<_>>();
+    let mut child = Command::new("cargo")
+        .args(&cmd_args)
+        .current_dir(&module_path)
+        .spawn()?;
+    child.wait()?;
+    let output_file = Path::new(&target)
+        .join("wasm32-wasi")
+        .join("debug")
         .join(name)
         .with_extension("wasm");
-    if !contract_build_path.exists() {
-        const TARGET_DIR_VAR: &str = "CARGO_TARGET_DIR";
-        let target = std::env::var(TARGET_DIR_VAR).map_err(|_| "CARGO_TARGET_DIR should be set")?;
-        println!("trying to compile the test contract, target: {target}");
-        // attempt to compile it
-        const RUST_TARGET_ARGS: &[&str] = &["build", "--target"];
-        const WASI_TARGET: &str = "wasm32-wasi";
-        let cmd_args = RUST_TARGET_ARGS
-            .iter()
-            .copied()
-            .chain([WASI_TARGET])
-            .collect::<Vec<_>>();
-        let mut child = Command::new("cargo")
-            .args(&cmd_args)
-            .current_dir(&module_path)
-            .spawn()?;
-        child.wait()?;
-        let output_file = Path::new(&target)
-            .join("wasm32-wasi")
-            .join("debug")
-            .join(name)
-            .with_extension("wasm");
-        println!("output file: {output_file:?}");
-        contract_build_path = output_file;
-    }
-    Ok(std::fs::read(contract_build_path)?)
+    println!("output file: {output_file:?}");
+    Ok(std::fs::read(output_file)?)
 }
 
-pub fn setup_test_contract(
+pub(crate) fn setup_test_contract(
     name: &str,
 ) -> Result<(ContractStore, ContractKey), Box<dyn std::error::Error>> {
-    let _ = tracing_subscriber::fmt().with_env_filter("info").try_init();
+    let _ = tracing_subscriber::fmt().with_env_filter("warn").try_init();
     let mut store = ContractStore::new(crate::tests::test_dir("contract"), 10_000)?;
     let contract_bytes = WrappedContract::new(
         Arc::new(ContractCode::from(get_test_module(name)?)),
